[FC] Don't require coding protocols on Inputs and Outputs
The particular coding necessary depends on if the Endpoint declaration
is going to be used for making or recieving requests. Instead of
requiring both on all Inputs and Outputs, I decided to remove the
requiremenet from the Endpoint protocols altogether.

Instead, the extensions that add makeRequest are the ones that require
the Input to be encodable and the Output to be decodable. The opposite
can be done when receiving requests.

The only downside that I see is that it could be confusing why
makeRequest isn't autocompleting.